refactor: 서비스-레포지토리 간 사용되는 dto entity로 변경
