+++ src/main/java/racingcar/dao/CarDao.java
@@ -4,9 +4,12 @@ package racingcar.dao;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
+import org.springframework.jdbc.core.namedparam.SqlParameterSourceUtils;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import racingcar.dto.CarDto;
+import racingcar.dao.entity.CarEntity;
+
+import java.util.List;
 
 @Repository
 public class CarDao {
@@ -20,8 +23,8 @@ public class CarDao {
                 .usingGeneratedKeyColumns("car_id");
     }
 
-    public int save(CarDto carDto) {
-        SqlParameterSource sqlParameterSource = new BeanPropertySqlParameterSource(carDto);
+    public int save(CarEntity carEntity) {
+        SqlParameterSource sqlParameterSource = new BeanPropertySqlParameterSource(carEntity);
         return insertActor.executeAndReturnKey(sqlParameterSource).intValue();
     }
 }
